requirements.txt
streamlit>=1.37
pandas>=2.0
pyyaml>=6.0
python-slugify>=8.0
